ssh-key: refactor encoding
Factors encoding support into separate `decoder` and `encoder` modules
