+++ ssh-key/src/decoder.rs
@@ -1,7 +1,6 @@
-//! Base64 support.
+//! Decoder support.
 //!
-//! Provides helper types which use the `base64ct` crate's constant-time Base64
-//! implementation for decoding.
+//! Support for decoding SSH keys from the OpenSSH wire format.
 
 use crate::{Error, Result};
 use core::str;
@@ -13,65 +12,17 @@ use alloc::{string::String, vec::Vec};
 /// Maximum size of a `usize` this library will accept.
 const MAX_SIZE: usize = 0xFFFFF;
 
-/// Get the estimated length of data when encoded as Base64.
-///
-/// This is an upper bound where the actual length might be slightly shorter.
-#[cfg(feature = "alloc")]
-pub(crate) fn encoded_len(input_len: usize) -> usize {
-    (((input_len * 4) / 3) + 3) & !3
-}
+/// Stateful Base64 decoder.
+pub(crate) type Base64Decoder<'i> = base64ct::Decoder<'i, base64ct::Base64>;
 
 /// Decoder trait.
 pub(crate) trait Decode: Sized {
     /// Attempt to decode a value of this type using the provided [`Decoder`].
-    fn decode(decoder: &mut impl DecoderExt) -> Result<Self>;
+    fn decode(decoder: &mut impl Decoder) -> Result<Self>;
 }
-
-/// Encoder trait.
-pub(crate) trait Encode: Sized {
-    /// Get the length of this type encoded in bytes, prior to Base64 encoding.
-    fn encoded_len(&self) -> Result<usize>;
-
-    /// Attempt to encode a value of this type using the provided [`Encoder`].
-    fn encode(&self, encoder: &mut impl EncoderExt) -> Result<()>;
-}
-
-/// String-like fields.
-///
-/// These fields receive a blanket impl of [`Decode`] and [`Encode`].
-pub(crate) trait StrField: AsRef<str> + str::FromStr<Err = Error> {
-    /// Decoding buffer type.
-    ///
-    /// This needs to be a byte array large enough to fit the largest
-    /// possible value of this type.
-    type DecodeBuf: AsMut<[u8]> + Default;
-}
-
-impl<T: StrField> Decode for T {
-    fn decode(decoder: &mut impl DecoderExt) -> Result<Self> {
-        let mut buf = T::DecodeBuf::default();
-        decoder.decode_str(buf.as_mut())?.parse()
-    }
-}
-
-impl<T: StrField> Encode for T {
-    fn encoded_len(&self) -> Result<usize> {
-        Ok(4 + self.as_ref().len())
-    }
-
-    fn encode(&self, encoder: &mut impl EncoderExt) -> Result<()> {
-        encoder.encode_str(self.as_ref())
-    }
-}
-
-/// Stateful Base64 decoder.
-pub(crate) type Decoder<'i> = base64ct::Decoder<'i, base64ct::Base64>;
-
-/// Stateful Base64 encoder.
-pub(crate) type Encoder<'o> = base64ct::Encoder<'o, base64ct::Base64>;
 
 /// Decoder extension trait.
-pub(crate) trait DecoderExt {
+pub(crate) trait Decoder {
     /// Decode as much Base64 as is needed to exactly fill `out`.
     ///
     /// This is the base decoding method on which the rest of the trait is
@@ -195,7 +146,7 @@ pub(crate) trait DecoderExt {
     }
 }
 
-impl DecoderExt for Decoder<'_> {
+impl Decoder for Base64Decoder<'_> {
     fn decode_base64<'o>(&mut self, out: &'o mut [u8]) -> Result<&'o [u8]> {
         Ok(self.decode(out)?)
     }
@@ -209,7 +160,7 @@ impl DecoderExt for Decoder<'_> {
     }
 }
 
-impl DecoderExt for pem::Decoder<'_> {
+impl Decoder for pem::Decoder<'_> {
     fn decode_base64<'o>(&mut self, out: &'o mut [u8]) -> Result<&'o [u8]> {
         Ok(self.decode(out)?)
     }
@@ -220,80 +171,5 @@ impl DecoderExt for pem::Decoder<'_> {
 
     fn is_finished(&self) -> bool {
         self.is_finished()
-    }
-}
-
-/// Encoder extension trait.
-pub(crate) trait EncoderExt {
-    /// Encode the given byte slice as Base64.
-    ///
-    /// This is the base encoding method on which the rest of the trait is
-    /// implemented in terms of.
-    fn encode_base64(&mut self, bytes: &[u8]) -> Result<()>;
-
-    /// Encode a `uint32` as described in [RFC4251 § 5]:
-    ///
-    /// > Represents a 32-bit unsigned integer.  Stored as four bytes in the
-    /// > order of decreasing significance (network byte order).
-    /// > For example: the value 699921578 (0x29b7f4aa) is stored as 29 b7 f4 aa.
-    ///
-    /// [RFC4251 § 5]: https://datatracker.ietf.org/doc/html/rfc4251#section-5
-    fn encode_u32(&mut self, num: u32) -> Result<()> {
-        self.encode_base64(&num.to_be_bytes())
-    }
-
-    /// Encode a `usize` as a `uint32` as described in [RFC4251 § 5].
-    ///
-    /// Uses [`Encoder::encode_u32`] after converting from a `usize`, handling
-    /// potential overflow if `usize` is bigger than `u32`.
-    ///
-    /// [RFC4251 § 5]: https://datatracker.ietf.org/doc/html/rfc4251#section-5
-    fn encode_usize(&mut self, num: usize) -> Result<()> {
-        self.encode_u32(u32::try_from(num)?)
-    }
-
-    /// Encodes `[u8]` into `byte[n]` as described in [RFC4251 § 5]:
-    ///
-    /// > A byte represents an arbitrary 8-bit value (octet).  Fixed length
-    /// > data is sometimes represented as an array of bytes, written
-    /// > byte[n], where n is the number of bytes in the array.
-    ///
-    /// [RFC4251 § 5]: https://datatracker.ietf.org/doc/html/rfc4251#section-5
-    fn encode_byte_slice(&mut self, bytes: &[u8]) -> Result<()> {
-        self.encode_usize(bytes.len())?;
-        self.encode_base64(bytes)
-    }
-
-    /// Encode a `string` as described in [RFC4251 § 5]:
-    ///
-    /// > Arbitrary length binary string.  Strings are allowed to contain
-    /// > arbitrary binary data, including null characters and 8-bit
-    /// > characters.  They are stored as a uint32 containing its length
-    /// > (number of bytes that follow) and zero (= empty string) or more
-    /// > bytes that are the value of the string.  Terminating null
-    /// > characters are not used.
-    /// >
-    /// > Strings are also used to store text.  In that case, US-ASCII is
-    /// > used for internal names, and ISO-10646 UTF-8 for text that might
-    /// > be displayed to the user.  The terminating null character SHOULD
-    /// > NOT normally be stored in the string.  For example: the US-ASCII
-    /// > string "testing" is represented as 00 00 00 07 t e s t i n g.  The
-    /// > UTF-8 mapping does not alter the encoding of US-ASCII characters.
-    ///
-    /// [RFC4251 § 5]: https://datatracker.ietf.org/doc/html/rfc4251#section-5
-    fn encode_str(&mut self, s: &str) -> Result<()> {
-        self.encode_byte_slice(s.as_bytes())
-    }
-}
-
-impl EncoderExt for Encoder<'_> {
-    fn encode_base64(&mut self, bytes: &[u8]) -> Result<()> {
-        Ok(self.encode(bytes)?)
-    }
-}
-
-impl EncoderExt for pem::Encoder<'_, '_> {
-    fn encode_base64(&mut self, bytes: &[u8]) -> Result<()> {
-        Ok(self.encode(bytes)?)
     }
 }
