+++ ssh-key/src/public/ecdsa.rs
@@ -1,7 +1,8 @@
 //! Elliptic Curve Digital Signature Algorithm (ECDSA) public keys.
 
 use crate::{
-    base64::{Decode, DecoderExt, Encode, EncoderExt},
+    decoder::{Decode, Decoder},
+    encoder::{Encode, Encoder},
     Algorithm, EcdsaCurve, Error, Result,
 };
 use core::fmt;
@@ -90,7 +91,7 @@ impl AsRef<[u8]> for EcdsaPublicKey {
 }
 
 impl Decode for EcdsaPublicKey {
-    fn decode(decoder: &mut impl DecoderExt) -> Result<Self> {
+    fn decode(decoder: &mut impl Decoder) -> Result<Self> {
         let curve = EcdsaCurve::decode(decoder)?;
 
         let mut buf = [0u8; Self::MAX_SIZE];
@@ -109,7 +110,7 @@ impl Encode for EcdsaPublicKey {
         Ok(4 + self.curve().encoded_len()? + self.as_ref().len())
     }
 
-    fn encode(&self, encoder: &mut impl EncoderExt) -> Result<()> {
+    fn encode(&self, encoder: &mut impl Encoder) -> Result<()> {
         self.curve().encode(encoder)?;
         encoder.encode_byte_slice(self.as_ref())
     }
