+++ ssh-key/src/public/rsa.rs
@@ -1,7 +1,8 @@
 //! Rivest–Shamir–Adleman (RSA) public keys.
 
 use crate::{
-    base64::{Decode, DecoderExt, Encode, EncoderExt},
+    decoder::{Decode, Decoder},
+    encoder::{Encode, Encoder},
     MPInt, Result,
 };
 
@@ -28,7 +29,7 @@ impl RsaPublicKey {
 }
 
 impl Decode for RsaPublicKey {
-    fn decode(decoder: &mut impl DecoderExt) -> Result<Self> {
+    fn decode(decoder: &mut impl Decoder) -> Result<Self> {
         let e = MPInt::decode(decoder)?;
         let n = MPInt::decode(decoder)?;
         Ok(Self { e, n })
@@ -40,7 +41,7 @@ impl Encode for RsaPublicKey {
         Ok(self.e.encoded_len()? + self.n.encoded_len()?)
     }
 
-    fn encode(&self, encoder: &mut impl EncoderExt) -> Result<()> {
+    fn encode(&self, encoder: &mut impl Encoder) -> Result<()> {
         self.e.encode(encoder)?;
         self.n.encode(encoder)
     }
